Add browser to integration test suite
