+++ lib/stream.js
@@ -1,6 +1,7 @@
 // Attempt to use readable-stream if available, attempt to use the built-in stream module.
 
 const ApiError = require("./error");
+const { streamAsyncIterator } = require("./util");
 const {
   EventSourceParserStream,
 } = require("../vendor/eventsource-parser/stream");
@@ -73,15 +74,7 @@ function createReadableStream({ url, fetch, options = {} }) {
         .pipeThrough(new TextDecoderStream())
         .pipeThrough(new EventSourceParserStream());
 
-      const reader = stream.getReader();
-
-      while (true) {
-        const { done, value: event } = await reader.read();
-
-        if (done) {
-          break;
-        }
-
+      for await (const event of streamAsyncIterator(stream)) {
         if (event.event === "error") {
           controller.error(new Error(event.data));
           break;
@@ -96,7 +89,6 @@ function createReadableStream({ url, fetch, options = {} }) {
         }
       }
 
-      reader.releaseLock();
       controller.close();
     },
   });
